refactor: extract critical_path into analyzer/critical_path.rs (#190)
Moves the Kahn-based longest-path computation out of audit() into a
standalone compute_critical_path(modules, dependencies) -> (max_depth,
critical_path) function. Tests move with it.
